ai/requirements.txt
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-deepseek>=0.1.0
langchain-community>=0.0.10
langchain-azure>=0.1.0
